Create script.js
JavaScript is used to create an instance of a new class of Bubble, passing
the circle properties. jQuery creates an html element with applied css properties
and common css class properties, this is prepended into the html body tag. 

A setTimeout JavaScript function is called for a timed display of bubble pops
where the draw method is used applying css properties to grow a circle and
further css properties to shrink the circle also a timed event.